Lots of little fixes to how links are created and accessed.
Changed folder structure so personal posts are in a specific folder and work posts are
in another. Removed a lot of dead links to external sites. Added Prev/Next navigation to
sidebar. Still work to be done there. Removed a lot of tags that were dups or not
necessary. Changed some styles on older posts to look better.